#7. Исправление замечаний (часть 2)
