--- src/view/form/offer-item-form-view.js
@@ -1,32 +1,0 @@
-import { createElement } from '../../render';
-
-
-const createOfferItemTemplate = () => `
-<div class="event__offer-selector">
-  <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked>
-  <label class="event__offer-label" for="event-offer-luggage-1">
-    <span class="event__offer-title">Add luggage</span>
-    &plus;&euro;&nbsp;
-    <span class="event__offer-price">30</span>
-  </label>
-</div>
-`;
-
-
-export default class OfferItemFormView {
-  getTemplate() {
-    return createOfferItemTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
