--- src/view/form/offers-form-view.js
@@ -1,30 +1,0 @@
-import { createElement } from '../../render';
-
-
-const createOffersFormTemplate = () => `
-  <section class="event__section  event__section--offers">
-    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-
-    <div class="event__available-offers">
-    </div>
-  </section>
-`;
-
-
-export default class OffersFormView {
-  getTemplate() {
-    return createOffersFormTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
